added host functions.

+++ libhost/osapi_posix.c
@@ -2,17 +2,17 @@
 #include "sys/time.h"
 #include "unistd.h"
 
-void _h_sleepmilli(long long milli) {
+void _h_sleepmilli(int64_t milli) {
     useconds_t micro = (useconds_t)milli * 1000;
     usleep(micro);
 }
 
-long long _h_nowmilli(void) {
+int64_t _h_nowmilli(void) {
     struct timeval time;
     gettimeofday(&time, NULL);
 
     return (
-        (long long)time.tv_sec  * 1000 +
-        (long long)time.tv_usec / 1000
+        (int64_t)time.tv_sec  * 1000 +
+        (int64_t)time.tv_usec / 1000
     );
 }
